+++ src/content/app/species-selector/views/species-selector-main-view/SpeciesSelectorMainView.tsx
@@ -33,7 +33,7 @@ const SpeciesSelectorMainView = () => {
   const onSearchSubmit = (query: string) => {
     trackSpeciesSearchQuery(query);
     navigate(
-      urlFor.speciesSelector({
+      urlFor.speciesSelectorSearch({
         query
       })
     );
